[ipv6] allow for translation func on forwarding to convert v6 to v4
In order to support DNS64, the interface for a backend now supports functions
for translating requests/responses that are forwarded to another DNS. The specific
implementation that we plan to create will convert AAAA record requests into A record
requests when going from ipv6 to ipv4. Then, the response to that request will be
translated so that the ipv4 ip is converted into an ipv6 ip and made into an AAAA record.